add deploy raft-consensus tools.
Former-commit-id: ba75f7009cc7f6063c1f3ae9a32fe40ec88b5cf8

+++ consensus/drivers/raft/tools/scripts/raft_deploy.sh
@@ -1,36 +1,30 @@
 #!/usr/bin/env bash
 #这是个用于分发部署chain33的脚本
+#Program:
+# This is a chain33 deploy scripts!
 SHELL_FOLDER=$(cd "$(dirname "$0")";pwd)
-echo "当前目录是:"$SHELL_FOLDER
-cd $SHELL_FOLDER/go-scp/
-go build -o go_scp
-cp go_scp  servers.toml ../
-rm -rf go_scp
-cd $SHELL_FOLDER
-./go_scp
-rm -rf go_scp
-rm -rf servers.toml
-rm -rf chain33.tgz
+echo "curl dir:"$SHELL_FOLDER
+if [ $1 == "start" ]
+then
+  cd $SHELL_FOLDER/go-scp/
+  go build -o go_scp
+  cp go_scp  servers.toml ../
+  rm -rf go_scp
+  cd $SHELL_FOLDER
+  ./go_scp start all
+#rm -rf go_scp
+#rm -rf servers.toml
+  rm -rf chain33.tgz
+elif [ $1 == "stop" ]
+then
+  ./go_scp stop all
+elif [ $1 == "clear" ]
+then
+   ./go_scp clear all
+else
+  echo "Usage: ./raft_deploy.sh [start,stop,clear]"
+fi
 
 
-#变量定义
-#ip_array=("raft15258.chinacloudapp.cn" )
-#user="ubuntu"
-#remote_cmd="ls /"
-#
-##本地通过ssh执行远程服务器的脚本
-#for ip in ${ip_array[*]}
-#do
-#    set timeout 30
-#    spawn ssh -l $user  $ip
-#    expect "password:"
-#    send "Fuzamei#123456\r"
-#    interact
-#    ls /
-##    if [ $ip = "192.168.1.1" ]; then
-##        port="7777"
-##    else
-##        port="22"
-##    fi
-##    ssh -t -p $port $user@$ip "remote_cmd"
-#done
+
+
